port(scared): real SCARED spiral geometry + float Node coords onto modern base
Ports the final Grandma-Millers SCARED mapping from the pre-modernization
master onto the LibGDX/Java21 branch. Replaces Brandon's early buildScared()
stub (2 Teensies, 192.168.0.x, linear node placement) with the real spiral
"christmas-tree" zome: 4 Teensies (192.168.50.x) × 8 strips × 200 LEDs, trig
node placement, 2 strips per tube, direct strip.nodeArray assignment (avoids
the array-doubling bug), and the 5th "Tesseract base" controller as commented
scaffolding.

- Node.x/y/z: int -> float, so trig/Sketchup positions keep sub-unit precision.
  Bounding-box fields (maxX..minZ) follow to float; TesseractApp particle->node
  mapping drops its (int) casts.
- Removed Processing helpers adapted to the render.ProcessingCompat shim
  (map, radians) and java.lang.Math (pow); TesseractMain.concat -> concatNodes.

Ports work from master commits ede0e38, 599c6be, 880a4ce, 0c0440f, f9e5edf.
Verified: headless boot with -DstageType=SCARED builds the stage
(maxW 835.9, maxH 842.9, maxD 592.5) and runs clips without error.

+++ src/main/app/TesseractApp.java
@@ -240,9 +240,9 @@ public class TesseractApp implements ApplicationListener, InputProcessor {
 
     private Node createNodeFromParticle(Particle p) {
         Node n = new Node();
-        n.x = (int) p.position.x;
-        n.y = (int) p.position.y;
-        n.z = (int) p.position.z;
+        n.x = p.position.x;
+        n.y = p.position.y;
+        n.z = p.position.z;
         n.r = (p.color >> 16) & 0xFF;
         n.g = (p.color >> 8) & 0xFF;
         n.b = p.color & 0xFF;
